tested N for X$ in weight items with or without limitation

+++ src/mainPackage/Inventory.java
@@ -37,7 +37,7 @@ public class Inventory {
 
     //NForXDollar: quantity params
     public void setNForXDollarSpecial(String name, int packageQuantity, BigDecimal packagePrice, boolean isSpecial){
-      setNForXDollarSpecial(name, packageQuantity, packagePrice, null, isSpecial);
+      setNForXDollarSpecial(name, packageQuantity, packagePrice, (Integer) null, isSpecial);
     }
 
     public void setNForXDollarSpecial(String name, int packageQuantity, BigDecimal packagePrice, Integer limit, boolean isSpecial){
@@ -50,7 +50,7 @@ public class Inventory {
     public void setNForXDollarSpecial(String name, double packageWeight, BigDecimal packagePrice, boolean isSpecial){
         setNForXDollarSpecial(name, packageWeight, packagePrice, null, isSpecial);
     }
-    public void setNForXDollarSpecial(String name, double packageWeight, BigDecimal packagePrice, Integer limit, boolean isSpecial){
+    public void setNForXDollarSpecial(String name, double packageWeight, BigDecimal packagePrice, Double limit, boolean isSpecial){
         Item item = findItem(name);
         item.setSpecial(isSpecial);
         item.setSpecialPrice(new NForXDollar(packageWeight, packagePrice, limit));
